Change the udf return type to float
to avoid long overflow.

+++ integration_tests/src/main/python/udf_test.py
@@ -77,8 +77,8 @@ def test_single_aggregate_udf(data_gen):
 @allow_non_gpu('AggregateInPandasExec', 'PythonUDF', 'Alias')
 @pytest.mark.parametrize('data_gen', integral_gens, ids=idfn)
 def test_group_aggregate_udf(data_gen):
-    @f.pandas_udf('long')
-    def pandas_sum(to_process: pd.Series) -> int:
+    @f.pandas_udf('double')
+    def pandas_sum(to_process: pd.Series) -> float:
         return to_process.sum()
 
     assert_gpu_and_cpu_are_equal_collect(
@@ -91,8 +91,8 @@ def test_group_aggregate_udf(data_gen):
 @allow_non_gpu('WindowInPandasExec', 'PythonUDF', 'WindowExpression', 'Alias', 'WindowSpecDefinition', 'SpecifiedWindowFrame', 'UnboundedPreceding$', 'UnboundedFollowing$')
 @pytest.mark.parametrize('data_gen', integral_gens, ids=idfn)
 def test_window_aggregate_udf(data_gen):
-    @f.pandas_udf('long')
-    def pandas_sum(to_process: pd.Series) -> int:
+    @f.pandas_udf('double')
+    def pandas_sum(to_process: pd.Series) -> float:
         return to_process.sum()
 
     w = Window\
